requirements.txt
dj-database-url==0.2.2
dj-static==0.0.5
Django==1.6.5
django-filter==0.10.0
django-hstore==1.4
djangorestframework==3.1.3
djrill==1.4.0
funcsigs==0.4
gunicorn==18.0
Markdown==2.6.2
mock==1.2.0
pbr==1.3.0
psycopg2==2.5.1
PyYAML==3.11
requests==2.7.0
six==1.9.0
South==0.8.4
static==0.4
